+++ Abs.h
@@ -6,8 +6,8 @@ public:
 	Abs();
 	~Abs();
 	double Execute() override { throw "Wrong arguments count"; };
-	double Execute(int x) override;
-	double Execute(int x, int y) override { throw "Wrong arguments count"; };
+	double Execute(double x) override;
+	double Execute(double x, double y) override { throw "Wrong arguments count"; };
 	
 };
