+++ Add.h
@@ -6,7 +6,7 @@ public:
 	Add();
 	~Add();
 	double Execute() override { throw "Wrong arguments count"; };
-	double Execute(int x) override { throw "Wrong arguments count"; };
-	double Execute(int x, int y) override;
+	double Execute(double x) override { throw "Wrong arguments count"; };
+	double Execute(double x, double y) override;
 };
